tixbot/requirements.txt
aiogram>=3.4,<4.0
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-dotenv>=1.0
python-multipart>=0.0.9
sqlalchemy>=2.0
asyncmy>=0.2.9
aiosqlite>=0.19
httpx>=0.27
